Remove styles for function parameters
Applies to c-like languages. This commit
removes the styles for function parameters
which were introduced in v1.0.4, since they
only worked for top-level functions and thus
provided an inconsistent user interface.

The problem is, that CodeMirror modes are
very unflexible when it comes to adding
more specific selectors to tokens.